+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Changelog
 Changelog for the dpdumper utility
 
+## [0.2.0] - 2024-08-04
+### Changed
+- Require dupicolib 0.3.2
+
+### Added
+- Added progress bar to track read/write completion
+- Added warning when using write mode
+- Added print of IC name when reading
+
 ## [0.1.2] - 2024-08-03
 ### Fixed
 - Fix execution in case the --definition parameter is not passed